corrected assignment of previous_picture in edit post

+++ vedrat/posts/routes.py
@@ -68,7 +68,10 @@ def postad(post_id=''):
 					post.description = form.description.data
 					post.posters_needed = form.posters.data
 					if post.image != 'default_ad_image.png':
+						delete_previous_picture = 'true'
 						previous_picture = post.image
+					else:
+						delete_previous_picture = 'false'
 					if form.image.data:
 						post.image = save_picture(form.image.data)
 					else:
@@ -78,7 +81,8 @@ def postad(post_id=''):
 					elif price_tag == 'add':
 						current_user.balance+=official_price
 					db.session.commit()
-					delete_picture(previous_picture)
+					if delete_previous_picture == 'true':
+						delete_picture(previous_picture)
 					flash('Your post has been updated successfully', 'success')
 					return redirect(url_for('users.userposts'))
 				else:
